remove vuex
