+++ package.json
@@ -25,8 +25,7 @@
     "typeface-roboto": "0.0.75",
     "typeface-rubik": "0.0.72",
     "uuid": "^8.3.0",
-    "vue": "^3.0.0",
-    "vuex": "^4.0.0-beta.4"
+    "vue": "^3.0.0"
   },
   "engine": "node >= 0.10",
   "devDependencies": {
